+++ src/cli/build.rs
@@ -4,9 +4,9 @@ use std::{
 };
 
 use anyhow::Context;
+use clap::Parser;
 use fs_err::File;
 use memofs::Vfs;
-use structopt::StructOpt;
 use tokio::runtime::Runtime;
 
 use crate::serve_session::ServeSession;
@@ -17,20 +17,20 @@ const UNKNOWN_OUTPUT_KIND_ERR: &str = "Could not detect what kind of file to bui
                                        Expected output file to end in .rbxl, .rbxlx, .rbxm, or .rbxmx.";
 
 /// Generates a model or place file from the Rojo project.
-#[derive(Debug, StructOpt)]
+#[derive(Debug, Parser)]
 pub struct BuildCommand {
     /// Path to the project to serve. Defaults to the current directory.
-    #[structopt(default_value = "")]
+    #[clap(default_value = "")]
     pub project: PathBuf,
 
     /// Where to output the result.
     ///
     /// Should end in .rbxm, .rbxl, .rbxmx, or .rbxlx.
-    #[structopt(long, short)]
+    #[clap(long, short)]
     pub output: PathBuf,
 
     /// Whether to automatically rebuild when any input files change.
-    #[structopt(long)]
+    #[clap(long)]
     pub watch: bool,
 }